fix: add package:version:create:list command, NUT
