+++ src/commands/force/package/beta/version/create/list.ts
@@ -8,6 +8,8 @@
 import * as os from 'os';
 import { flags, FlagsConfig, SfdxCommand } from '@salesforce/command';
 import { Messages } from '@salesforce/core';
+import { PackageVersion, PackageVersionCreateRequestResult } from '@salesforce/packaging';
+import * as chalk from 'chalk';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-packaging', 'package_version_create_list');
@@ -31,8 +33,42 @@ export class PackageVersionCreateListCommand extends SfdxCommand {
     }),
   };
 
-  // eslint-disable-next-line @typescript-eslint/require-await
-  public async run(): Promise<never> {
-    throw new Error('Beta command not yet implemented');
+  public async run(): Promise<PackageVersionCreateRequestResult[]> {
+    const pv = new PackageVersion({ connection: this.hubOrg.getConnection(), project: undefined });
+    const results = await pv.createdList({ ...this.flags });
+
+    if (results.length === 0) {
+      this.ux.log('No results found');
+    } else {
+      this.ux.styledHeader(chalk.blue(`Package Version Create Requests  [${results.length}]`));
+      const columnData = {
+        Id: {},
+        Status: {
+          header: messages.getMessage('status'),
+        },
+        Package2Id: {
+          header: messages.getMessage('packageId'),
+        },
+        Package2VersionId: {
+          header: messages.getMessage('packageVersionId'),
+        },
+        SubscriberPackageVersionId: {
+          header: messages.getMessage('subscriberPackageVersionId'),
+        },
+        Tag: {
+          header: messages.getMessage('tag'),
+        },
+        Branch: {
+          header: messages.getMessage('branch'),
+        },
+        CreatedDate: { header: 'Created Date' },
+        CreatedBy: {
+          header: messages.getMessage('createdBy'),
+        },
+      };
+      this.ux.table(results, columnData, { 'no-truncate': true });
+    }
+
+    return results;
   }
 }
